+++ junebug/channel.py
@@ -101,12 +101,22 @@ class Channel(object):
     def character_limit(self):
         return self._properties.get('character_limit')
 
+    @property
+    def has_destination(self):
+        """
+        Whether or not this channel has a message destination defined
+        """
+        return 'mo_url' in self._properties or 'amqp_queue' in self._properties
+
     @inlineCallbacks
     def start(self, service, transport_worker=None):
         '''Starts the relevant workers for the channel. ``service`` is the
         parent of under which the workers should be started.'''
         self._start_transport(service, transport_worker)
-        self._start_application(service)
+        # Only start the application worker if we have somewhere to send the
+        # messages.
+        if self.has_destination:
+            self._start_application(service)
         self._start_status_application(service)
         for plugin in self.plugins:
             yield plugin.channel_started(self)
@@ -244,13 +254,10 @@ class Channel(object):
     @inlineCallbacks
     def send_message(self, sender, outbounds, msg):
         '''Sends a message.'''
-        event_url = msg.get('event_url')
-        event_auth_token = msg.get('event_auth_token', None)
-        msg = message_from_api(self.id, msg)
-        msg = TransportUserMessage.send(**msg)
-        msg = yield self._send_message(sender, outbounds, event_url, msg,
-                                       event_auth_token)
-        returnValue(api_from_message(msg))
+        vumi_msg = message_from_api(self.id, msg)
+        vumi_msg = TransportUserMessage.send(**vumi_msg)
+        vumi_msg = yield self._send_message(sender, outbounds, vumi_msg, msg)
+        returnValue(api_from_message(vumi_msg))
 
     @inlineCallbacks
     def send_reply_message(self, sender, outbounds, inbounds, msg,
@@ -268,13 +275,10 @@ class Channel(object):
             raise MessageNotFound(
                 "Inbound message with id %s not found" % (msg['reply_to'],))
 
-        event_url = msg.get('event_url')
-        event_auth_token = msg.get('event_auth_token', None)
-        msg = message_from_api(self.id, msg)
-        msg = in_msg.reply(**msg)
-        msg = yield self._send_message(sender, outbounds, event_url, msg,
-                                       event_auth_token)
-        returnValue(api_from_message(msg))
+        vumi_msg = message_from_api(self.id, msg)
+        vumi_msg = in_msg.reply(**vumi_msg)
+        vumi_msg = yield self._send_message(sender, outbounds, vumi_msg, msg)
+        returnValue(api_from_message(vumi_msg))
 
     def get_logs(self, n):
         '''Returns the last `n` logs. If `n` is greater than the configured
@@ -412,7 +416,12 @@ class Channel(object):
 
     def _restore(self, service):
         self.transport_worker = service.getServiceNamed(self.id)
-        self.application_worker = service.getServiceNamed(self.application_id)
+        try:
+            self.application_worker = service.getServiceNamed(
+                self.application_id)
+        except KeyError:
+            # Doesn't have an application worker if no destination specified
+            pass
         self.status_application_worker = service.getServiceNamed(
             self.status_application_id)
 
@@ -426,17 +435,11 @@ class Channel(object):
                 )
 
     @inlineCallbacks
-    def _send_message(self, sender, outbounds, event_url, msg,
-                      event_auth_token=None):
+    def _send_message(self, sender, outbounds, msg, msg_api):
         self._check_character_limit(msg['content'])
 
-        if event_url is not None:
-            yield outbounds.store_event_url(
-                self.id, msg['message_id'], event_url)
-            if event_auth_token is not None:
-                yield outbounds.store_event_auth_token(
-                    self.id, msg['message_id'], event_auth_token
-                )
+        msg_api.update(api_from_message(msg))
+        yield outbounds.store_message(self.id, msg_api)
 
         queue = self.OUTBOUND_QUEUE % (self.id,)
         msg = yield sender.send_message(msg, routing_key=queue)